main.py
#!/usr/bin/env python3
"""
EvalAI Challenge #2508 - Main Implementation
"""

import os
import sys
import logging
from pathlib import Path
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

class EvalAIChallenge:
    """Main class for handling the EvalAI challenge."""
    
    def __init__(self):
        self.challenge_id = 2508
        self.challenge_url = "https://eval.ai/web/challenges/challenge-page/2508/overview"
        
    def setup(self):
        """Initialize the challenge environment."""
        logger.info(f"Setting up EvalAI Challenge #{self.challenge_id}")
        # TODO: Add setup logic based on challenge requirements
        
    def load_data(self):
        """Load challenge data."""
        logger.info("Loading challenge data...")
        # TODO: Implement data loading based on challenge requirements
        
    def preprocess_data(self):
        """Preprocess the data for the challenge."""
        logger.info("Preprocessing data...")
        # TODO: Implement data preprocessing
        
    def train_model(self):
        """Train the model for the challenge."""
        logger.info("Training model...")
        # TODO: Implement model training
        
    def evaluate_model(self):
        """Evaluate the model performance."""
        logger.info("Evaluating model...")
        # TODO: Implement model evaluation
        
    def generate_submission(self):
        """Generate submission file for EvalAI."""
        logger.info("Generating submission...")
        # TODO: Implement submission generation
        
    def run(self):
        """Run the complete challenge pipeline."""
        logger.info("Starting EvalAI Challenge pipeline...")
        
        try:
            self.setup()
            self.load_data()
            self.preprocess_data()
            self.train_model()
            self.evaluate_model()
            self.generate_submission()
            
            logger.info("Challenge pipeline completed successfully!")
            
        except Exception as e:
            logger.error(f"Error in challenge pipeline: {e}")
            raise

def main():
    """Main entry point."""
    challenge = EvalAIChallenge()
    challenge.run()

if __name__ == "__main__":
    main() 