requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep learning (if needed)
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0

# Data processing
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0
click>=8.0.0

# Jupyter (for development)
jupyter>=1.0.0
ipykernel>=6.0.0 